Rename People Search and Quick Search
