+++ src/views/ApartmentApp/components/StudentApplication/components/ApplicantList/components/ApplicantListItem/index.js
@@ -14,7 +14,7 @@ import StarIcon from '@material-ui/icons/Star';
 import StarBorderIcon from '@material-ui/icons/StarBorder';
 import { Fragment, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
-import { Class } from 'services/goStalk';
+import { Class } from 'services/peopleSearch';
 import user from 'services/user';
 import styles from '../../../../../../ApartmentApp.module.css';
